server/requirements.txt
fastapi
uvicorn
beanie
motor
pydantic
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
PyMuPDF
pdfminer.six
openai
httpx 